move unpack and pack functions to ethexe/runtime/common
