+++ ethexe/processor/src/host/api/sandbox.rs
@@ -20,7 +20,7 @@
 
 use crate::host::{api::MemoryWrap, context::HostContext};
 use anyhow::Result;
-use ethexe_common::pack_u32_to_i64;
+use ethexe_runtime_common::pack_u32_to_i64;
 use gear_runtime_interface::{sandbox_detail, Instantiate};
 use parity_scale_codec::Encode;
 use sp_wasm_interface::{FunctionContext as _, IntoValue as _, Pointer, StoreData};
