+++ ethexe/runtime/common/src/lib.rs
@@ -270,3 +270,17 @@ where
     core_processor::process::<Ext<RI::LazyPages>>(&block_config, execution_context, random_data)
         .unwrap_or_else(|err| unreachable!("{err}"))
 }
+
+pub const fn pack_u32_to_i64(low: u32, high: u32) -> i64 {
+    let mut result = 0u64;
+    result |= (high as u64) << 32;
+    result |= low as u64;
+    result as i64
+}
+
+pub const fn unpack_i64_to_u32(val: i64) -> (u32, u32) {
+    let val = val as u64;
+    let high = (val >> 32) as u32;
+    let low = val as u32;
+    (low, high)
+}
